fix: iterate on review for #28
