+++ crates/beautiful-mermaid-rs/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "beautiful-mermaid-rs"
+version = "0.0.0"
+edition = "2021"
+
+[lib]
+path = "src/ascii/shapes/stadium.rs"
