+++ crates/beautiful-mermaid-rs/src/ascii/shapes/stadium.rs
@@ -41,6 +41,7 @@ pub enum Direction {
 }
 
 pub type Canvas = Vec<Vec<char>>;
+mod rectangle;
 
 pub fn get_dimensions(label: &str, options: ShapeRenderOptions) -> ShapeDimensions {
     let lines = split_lines(label);
@@ -149,61 +150,6 @@ fn split_lines(label: &str) -> Vec<&str> {
 
 fn code_point_width(line: &str) -> usize {
     line.chars().count()
-}
-
-mod rectangle {
-    use super::{Direction, DrawingCoord, ShapeDimensions};
-
-    pub fn get_box_attachment_point(
-        dir: Direction,
-        dimensions: &ShapeDimensions,
-        base_coord: DrawingCoord,
-    ) -> DrawingCoord {
-        let width = dimensions.width as isize;
-        let height = dimensions.height as isize;
-
-        let center_x = base_coord.x + (width / 2);
-        let center_y = base_coord.y + (height / 2);
-
-        match dir {
-            Direction::Up => DrawingCoord {
-                x: center_x,
-                y: base_coord.y,
-            },
-            Direction::Down => DrawingCoord {
-                x: center_x,
-                y: base_coord.y + height - 1,
-            },
-            Direction::Left => DrawingCoord {
-                x: base_coord.x,
-                y: center_y,
-            },
-            Direction::Right => DrawingCoord {
-                x: base_coord.x + width - 1,
-                y: center_y,
-            },
-            Direction::UpperLeft => DrawingCoord {
-                x: base_coord.x,
-                y: base_coord.y,
-            },
-            Direction::UpperRight => DrawingCoord {
-                x: base_coord.x + width - 1,
-                y: base_coord.y,
-            },
-            Direction::LowerLeft => DrawingCoord {
-                x: base_coord.x,
-                y: base_coord.y + height - 1,
-            },
-            Direction::LowerRight => DrawingCoord {
-                x: base_coord.x + width - 1,
-                y: base_coord.y + height - 1,
-            },
-            Direction::Middle => DrawingCoord {
-                x: center_x,
-                y: center_y,
-            },
-        }
-    }
 }
 
 #[cfg(test)]
